Updates 2022-03-07 notes
Just need to update templates

✅darkened footer and news/announcement pages further
✅for now, try mönke (err on making reading easier), behind green bars, color

ARGUMENT: with the green boxes going all the way across, and with a much more colorful image beneath, id argue for either a. reducing the bars back down instead of spreading all the way across or b. changing the color to like a dark gray so it doesn't muddy up monkë's colors, but just my opinion